change logger init
